+++ src/mainwindow.cpp
@@ -187,6 +187,9 @@ void MainWindow::updateChargingStatus()
         case battery_fully_charged:
             chargingStatus = tr("Fully charged");
             break;
+        case battery_fully_charged_no_power:
+            chargingStatus = tr("Fully charged (Discharging)");
+            break;
         default:
             chargingStatus = tr("Unknown");
             break;